Add support for data:text/tml urls
Added Support for inline iframe content. This allows urls of type "data:text/html" to be specified in the URL field. The HTML of the iframe then follows in either URL-encoded format or base64 econded format.

For example, the url specified as 
```
   data:text/html;charset=utf-8;base64, PGh0bWw+PGJvZHk+SGVsbG8gV29ybGQ8L2JvZHk+PC9odG1sPg==
```
Will add an iframe with the "hello world" text in it.